Ranger mirage line

+++ fish/config.fish
@@ -318,13 +318,25 @@ end
 tabs -2
 
 function ranger-cd
-  set tempfile "/tmp/(whoami)chosendir"
-  ranger --choosedir=$tempfile (pwd)
-  if test -f $tempfile
-      if [ (cat $tempfile) != (pwd) ]
-        cd (cat $tempfile)
-      end
-  end
-  rm -f $tempfile
+	if test -z "$RANGER_LEVEL"
+		set tempfile "/tmp/(whoami)chosendir"
+		ranger --choosedir=$tempfile (pwd)
+		if test -f $tempfile
+				if [ (cat $tempfile) != (pwd) ]
+					cd (cat $tempfile)
+				end
+		end
+		rm -f $tempfile
+	else
+		exit
+	end
 end
 alias ll='ranger-cd'
+
+if test -n "$RANGER_LEVEL"
+	clear
+	echo "===================="
+	echo "=   RANGER SHELL   ="
+	echo "===================="
+	ls -l
+end
